+++ src/index.ts
@@ -19,6 +19,8 @@ export { Inventory } from './components/Inventory'
 export type { InventoryProps } from './components/Inventory'
 export { Shop } from './components/Shop'
 export type { ShopProps } from './components/Shop'
+export { Overlay } from './components/Overlay'
+export type { OverlayProps } from './components/Overlay'
 export { useNoragonKeyboard } from './components/useNoragonKeyboard'
 export type { UseNoragonKeyboardOptions } from './components/useNoragonKeyboard'